Refactor the IO::AudioFile namespace into headers and sources
