+++ includes/IO/AudioFile/lib/libsndfile_Read.cpp
@@ -1,5 +1,3 @@
-#pragma once
-
 /*
  * /////////
  * // Clover
@@ -21,18 +19,46 @@
  */
 
 #include <stdexcept>
-#include <string>
-#include <vector>
 
 #include "sndfile.h"
 
 #include "../AudioFile.h"
-#include "../AudioFileWriteSettings.h"
 #include "libsndfile__Util.h"
 
 namespace Clover::IO::AudioFile::impl {
 
-void libsndfile_Write(const char *path, const WriteSettingsPcm &writeSettings,
-                      const AudioFile &audioFile);
+AudioFile libsndfile_Read(const char *path) {
+  AudioFile audioFile;
+  SF_INFO sfinfo;
+  SNDFILE *file = sf_open(path, SFM_READ, &sfinfo);
+  throwIfFails(file, sf_error(file));
+
+  audioFile.sampleRateHz = sfinfo.samplerate;
+  audioFile.channelCount = sfinfo.channels;
+
+  audioFile.audioData.resize(sfinfo.frames * sfinfo.channels);
+  sf_count_t count = sf_read_float(file, audioFile.audioData.data(),
+                                   audioFile.audioData.size());
+
+  if (count != static_cast<sf_count_t>(audioFile.audioData.size())) {
+    throwIfFails(file, sf_error(file));
+  }
+
+  SF_CUES cues;
+  if (sf_command(file, SFC_GET_CUE, &cues, sizeof(cues)) == SF_TRUE) {
+    for (unsigned i = 0; i < cues.cue_count; ++i) {
+      audioFile.cuePoints.push_back(cues.cue_points[i].sample_offset);
+    }
+  }
+
+  if (sf_close(file) != 0) {
+    int err = sf_error(file);
+    if (err != SF_ERR_NO_ERROR) {
+      throw std::runtime_error(sf_strerror(file));
+    }
+  }
+
+  return audioFile;
+}
 
 } // namespace Clover::IO::AudioFile::impl
