+++ includes/IO/AudioFile/lib/libsndfile__Util.h
@@ -20,67 +20,13 @@
  *
  */
 
-#include <stdexcept>
-#include <string>
-#include <vector>
-
 #include "sndfile.h"
 
-#include "../AudioFile.h"
 #include "../AudioFileWriteSettings.h"
 
 namespace Clover::IO::AudioFile::impl {
 
-void throwIfFails(SNDFILE *sndfile, int resultCode) {
-  bool hasError = resultCode != SF_ERR_NO_ERROR;
-  bool hasFile = !!sndfile;
-
-  const char *errorMessage = sf_strerror(sndfile);
-
-  if (hasFile && hasError) {
-    sf_close(sndfile);
-  }
-  if (hasError) {
-    throw std::runtime_error(errorMessage);
-  }
-}
-
-int getWriteSettingsFormat(const WriteSettingsPcm &writeSettings) {
-  int flags = 0;
-
-  int bitDepth = 0;
-
-  switch (writeSettings.bitDepth) {
-  case PcmBitDepth::_16:
-    flags |= SF_FORMAT_PCM_16;
-    break;
-  case PcmBitDepth::_24:
-    flags |= SF_FORMAT_PCM_24;
-    break;
-  case PcmBitDepth::_32:
-    flags |= SF_FORMAT_PCM_32;
-    break;
-  case PcmBitDepth::_float:
-    flags |= SF_FORMAT_FLOAT;
-    break;
-  case PcmBitDepth::_double:
-    flags |= SF_FORMAT_DOUBLE;
-    break;
-  default:
-    flags |= SF_FORMAT_PCM_16;
-    break;
-  }
-
-  switch (writeSettings.pcmFileType) {
-  case PcmFileType::Wav:
-    flags |= SF_FORMAT_WAV;
-    break;
-  case PcmFileType::Flac:
-    flags |= SF_FORMAT_FLAC;
-    break;
-  }
-
-  return flags;
-}
+void throwIfFails(SNDFILE *sndfile, int resultCode);
+int getWriteSettingsFormat(const WriteSettingsPcm &writeSettings);
 
 } // namespace Clover::IO::AudioFile::impl
